+++ Framework/MobileDevice/Device Listener/SDMMD_USBMuxListener.h
@@ -37,7 +37,7 @@
 #pragma mark Functions
 #pragma mark -
 
-sdmmd_return_t SDMMD_USBMuxConnectByPort(SDMMD_AMDeviceRef device, uint32_t port, uint32_t *socketConn);
+sdmmd_return_t SDMMD_USBMuxConnectByPort(SDMMD_AMDeviceRef device, uint16_t port, uint32_t *socketConn);
 
 SDMMD_USBMuxListenerRef SDMMD_USBMuxCreate();
 void SDMMD_USBMuxStartListener(SDMMD_USBMuxListenerRef *listener);
